feat: add conditional rendering for MainPage
