+++ src/components/Header/Header.module.scss
@@ -7,6 +7,7 @@
   align-items: center;
   justify-content: space-between;
   height: 100px;
+  padding: 0 40px;
   background: white;
   transition: all 0.3s ease;
 }
