+++ src/store/userStore.ts
@@ -2,18 +2,19 @@ import { create } from 'zustand';
 import { persist } from 'zustand/middleware';
 
 interface UserState {
-  user: boolean;
-  setUser: (value: boolean) => void;
-  toggleUser: () => void;
+  user: null | { uid: string; email?: string };
+  setUser: (user: UserState['user']) => void;
 }
 
 export const useUserStore = create<UserState>()(
   persist(
     (set) => ({
-      user: false,
-      setUser: (value: boolean) => set({ user: value }),
-      toggleUser: () => set((state) => ({ user: !state.user })),
+      user: null,
+      setUser: (user) => set({ user }),
     }),
-    { name: 'user' }
+    {
+      name: 'user',
+      partialize: (state) => ({ user: state.user }),
+    }
   )
 );
